Fuzzy Flocking Initial Commit
The plan is to illustrate a flocking pattern where by "birds" fly in a 2D plane and interact with each other in a fashion similar to "flocking". The intention is to apply fuzzy logic as a controller system.

The initial commit includes basic mechanics for geometry (Vector based arithmetic) and an environment in which newtonian-ish physics can be applied to the birds. As well, the basic execution of a fuzzy logic system applying acceleration to the objects is in place. Proper rules need to be defined.